+++ backend/src/routes/v2/model/getModelsSearch.ts
@@ -64,6 +64,7 @@ export interface ModelSearchResult {
   collaborators: Array<CollaboratorEntry>
   createdAt: Date
   updatedAt: Date
+  sourceModelId?: string
 }
 
 interface GetModelsResponse {
@@ -98,8 +99,10 @@ export const getModelsSearch = [
       organisation: model.organisation,
       state: model.state,
       collaborators: model.collaborators,
+      visibility: model.visibility,
       createdAt: model.createdAt,
       updatedAt: model.updatedAt,
+      sourceModelId: model.settings?.mirror?.sourceModelId,
     }))
 
     await audit.onSearchModel(req, models)