+++ frontend/pages/model/[modelId].tsx
@@ -21,7 +21,7 @@ import { getCurrentUserRoles } from 'utils/roles'
 export default function Model() {
   const router = useRouter()
   const { modelId }: { modelId?: string } = router.query
-  const { model, isModelLoading, isModelError } = useGetModel(modelId, EntryKind.MODEL)
+  const { model, isModelLoading, isModelError } = useGetModel(modelId)
   const { currentUser, isCurrentUserLoading, isCurrentUserError } = useGetCurrentUser()
   const { uiConfig, isUiConfigLoading, isUiConfigError } = useGetUiConfig()
 
@@ -38,7 +38,7 @@ export default function Model() {
             {
               title: 'Overview',
               path: 'overview',
-              view: <Overview entry={model} readOnly={!!model.settings.mirror?.sourceModelId} />,
+              view: <Overview entry={model} readOnly={model.kind === EntryKind.MIRRORED_MODEL} />,
             },
             {
               title: 'Releases',
@@ -117,7 +117,6 @@ export default function Model() {
           requiredUrlParams={{ modelId: model.id }}
           titleToCopy={model.name}
           subheadingToCopy={model.id}
-          sourceModelId={model.settings.mirror?.sourceModelId}
         />
       )}
     </>