+++ frontend/src/MessageAlert.tsx
@@ -7,7 +7,7 @@ import { useEffect, useRef, useState } from 'react'
 import CopyToClipboardButton from 'src/common/CopyToClipboardButton'
 import Link from 'src/Link'
 
-type PartialMessageAlertProps =
+type PartialLinkMessageAlertProps =
   | {
       linkText: string
       href: string
@@ -17,31 +17,46 @@ type PartialMessageAlertProps =
       href?: never
     }
 
+type PartialButtonMessageAlertProps =
+  | {
+      buttonText: string
+      buttonAction: () => void
+    }
+  | {
+      buttonText?: never
+      buttonAction?: never
+    }
+
 type MessageAlertProps = {
   message?: string
   severity?: AlertProps['severity']
   'data-test'?: string
   slimView?: boolean
-} & PartialMessageAlertProps
+  disableScrollToView?: boolean
+} & PartialLinkMessageAlertProps &
+  PartialButtonMessageAlertProps
 
 export default function MessageAlert({
   message = '',
   severity,
   linkText,
   href,
+  buttonText,
+  buttonAction,
   'data-test': dataTest,
   slimView = false,
+  disableScrollToView = false,
 }: MessageAlertProps) {
   const alertRef = useRef<HTMLDivElement>(null)
   const [showContactMessage, setShowContactMessage] = useState(false)
 
   useEffect(() => {
-    if (message && alertRef.current && alertRef.current.scrollIntoView) {
+    if (!disableScrollToView && message && alertRef.current && alertRef.current.scrollIntoView) {
       alertRef.current.scrollIntoView({
         behavior: 'smooth',
       })
     }
-  }, [message])
+  }, [message, disableScrollToView])
 
   const handleContact = () => {
     setShowContactMessage(!showContactMessage)
@@ -70,6 +85,7 @@ export default function MessageAlert({
               ariaLabel='copy error message to clipboard'
             />
           )}
+          {buttonText && <Button onClick={buttonAction}>{buttonText}</Button>}
         </Stack>
         {!!(href && linkText) && (
           <Typography>