+++ frontend/src/marketplace/EntryListRow.tsx
@@ -1,10 +1,11 @@
 import { CorporateFare } from '@mui/icons-material'
-import { Box, Divider, Stack, Typography } from '@mui/material'
+import { Box, Chip, Divider, Stack, Typography } from '@mui/material'
 import { useTheme } from '@mui/material/styles'
 import { EntrySearchResult } from 'actions/model'
-import { CSSProperties } from 'react'
+import { CSSProperties, useMemo } from 'react'
 import ChipSelector from 'src/common/ChipSelector'
 import Link from 'src/Link'
+import { EntryKind } from 'types/types'
 
 interface EntryListRowProps {
   selectedChips: string[]
@@ -36,6 +37,16 @@ export default function EntryListRow({
   const theme = useTheme()
   const entry = data[index]
 
+  const entryKindForRedirect = useMemo(() => {
+    return entry.kind === EntryKind.MODEL || entry.kind === EntryKind.MIRRORED_MODEL ? EntryKind.MODEL : entry.kind
+  }, [entry])
+
+  const mirroredLabel = useMemo(() => {
+    if (entry.kind === EntryKind.MIRRORED_MODEL) {
+      return <Typography>Mirrored</Typography>
+    }
+  }, [entry])
+
   return (
     <Box
       justifyContent='flex-start'
@@ -51,21 +62,29 @@ export default function EntryListRow({
       <Stack spacing={1}>
         <Link
           sx={{ textDecoration: 'none', whiteSpace: 'nowrap', textOverflow: 'ellipsis', overflow: 'hidden' }}
-          href={`${entry.kind}/${entry.id}`}
+          href={`${entryKindForRedirect}/${entry.id}`}
         >
-          <Typography
-            variant='h5'
-            component='h2'
-            sx={{
-              fontWeight: '500',
-              textDecoration: 'none',
-              color: theme.palette.primary.main,
-              textOverflow: 'ellipsis',
-              overflow: 'hidden',
-            }}
-          >
-            {entry.name}
-          </Typography>
+          <Stack spacing={1} justifyContent='space-between' alignItems='center' direction='row'>
+            <Typography
+              variant='h5'
+              component='h2'
+              sx={{
+                fontWeight: '500',
+                textDecoration: 'none',
+                color: theme.palette.primary.main,
+                textOverflow: 'ellipsis',
+                overflow: 'hidden',
+              }}
+            >
+              {entry.name}
+            </Typography>
+            <Stack spacing={2} direction='row'>
+              {entry.kind === EntryKind.MIRRORED_MODEL && (
+                <Chip size='small' color='secondary' variant='outlined' label={mirroredLabel} />
+              )}
+              {entry.visibility === 'private' && <Chip size='small' color='secondary' label='Private' />}
+            </Stack>
+          </Stack>
         </Link>
         <Typography variant='body1' sx={{ whiteSpace: 'nowrap', textOverflow: 'ellipsis', overflow: 'hidden' }}>
           {entry.description}