+++ lib/python/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All dates are formatted dd/mm/yyyy.
 
-## 3.2.0 - 20/10/2025
+## 3.2.2 - 20/10/2025
 
 ### Breaking Changes
 
@@ -13,9 +13,15 @@ All dates are formatted dd/mm/yyyy.
 - Add Python 3.14 support.
 - Update package dependencies.
 
-## 3.1.2 - 30/09/2025
+## 3.2.1 30/10/2025
 
-### Changes
+- Add `tags` property to `Entry` (and inherited classes `Model`, `MirroredModel` and `Datacard`)
+
+## 3.2.0 29/10/2025
+
+- Added new `MirroredModel` class.
+
+## 3.1.2 30/09/2025
 
 - Fix broken `Client.put_file_scan` method.
 - Update package dependencies.